Fix build Gcc

+++ Game/src/Systems/PlayerCameraSystem.cpp
@@ -89,7 +89,7 @@ void PlayerCameraSystem::Update(ThreadContext& threadContext, Vault& vault) {
 	// Update tracked rotation and zoom
 	camera.rotation += requestedRotation * ROTATION_SPEED * gTime.deltaTime;
 	camera.rotation.y = std::clamp(camera.rotation.y, -50.f, 50.f);
-    camera.rotation.x = std::fmodf(camera.rotation.x, 360.f);
+    camera.rotation.x = std::fmod(camera.rotation.x, 360.f);
 	camera.zoom = std::clamp(camera.zoom - requestedZoom * ZOOM_SPEED * gTime.deltaTime, 0.2f, 2.f);
 
 	// Compute new camera location and orientation based on the rotation and zoom
@@ -99,7 +99,7 @@ void PlayerCameraSystem::Update(ThreadContext& threadContext, Vault& vault) {
     float playerYaw = DegreesToRadians(-camera.rotation.x);
     if (camera.directionFromPlayer) {
         const Vec3 playerForward = interpolatedRotation.ToRotationMatrix() * -Vec3::FORWARD;
-        playerYaw = std::atan2f(playerForward.x, playerForward.z);
+        playerYaw = std::atan2(playerForward.x, playerForward.z);
     }
 
 	const Quat cameraYawRotation = Quat(playerYaw, yawAxis);
